Refactor logging, DI, and texture decoding; modernize code
- Switched to non-namespaced ILogger and modern DI patterns throughout
- Split and improved IAudioConverter, IMediaProcessor, ITextureService interfaces
- Enhanced DefaultTextureService with robust image/file conversion
- Refactored UbiArt/Unity format classes for new service patterns
- Modernized code style: explicit types, collection expressions, validation
- Fixed GTX texture decoding: BCn handling, mip offsets, address math
- Updated tests for clarity and type safety; fixed minor bugs
- Modernized UI/console dialogues and improved error handling
- General cleanup: removed dead code, improved readability and maintainability

+++ JustDanceEditor.Editor/Views/Timeline/AudioBarControl.cs
@@ -212,7 +212,7 @@ public class AudioBarControl : Control
             int step = Math.Max(1, Samples.Length / (int)Math.Max(1, totalWidth));
             for (int x = 0; x < (int)totalWidth; x++)
             {
-                int sampleIdx = (int)((double)x / totalWidth * Samples.Length);
+                int sampleIdx = (int)(x / totalWidth * Samples.Length);
                 if (sampleIdx >= Samples.Length)
                     break;
